Ajustes e acertos globais, routes e estrutura projeto
