+++ labeddit/src/components/CreatePost/styles.js
@@ -1,0 +1,6 @@
+import styled from 'styled-components';
+
+export const PostPageContainer = styled.div `
+
+    width: 99%;
+`
