--- labeddit/src/pages/PostPage/styles.js
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const PostPageContainer = styled.div `
-
-    width: 99%;
-`
